Avoid deadlocking in profiler when `end_heap_allocation_recording` raises
Due to a bug in the integration between how the heap sampler works and
the new "missing allocations/skipped samples" placeholder, we could
cause one of the heap sampler sanity checks to fail, and thus raise
an exception.

This exception would be correctly logged, e.g. something like
> W, [2024-07-08T11:24:21.212654 #226523]  WARN -- datadog: [datadog]
> (lib/datadog/profiling/collectors/cpu_and_wall_time_worker.rb:72:in
> `rescue in block (2 levels) in start') CpuAndWallTimeWorker thread
> error. Cause: RuntimeError Ended a heap recording that was not
> started Location: missing-testcase.rb:2:in `new'

The exception would also cause the profiler to try to shut down.

But this is where we would run into issues: the scheduler still tries
to report one last profile before shutting down, and because the
above exception left the stack recorder with the active slot mutex
locked, this would make the scheduler deadlock forever, thus
preventing the Ruby VM from shutting down.

Part 1 on fixing this issue is to correctly unlock the mutex, thus
getting rid of the deadlock.

Part 2 (next commit) will be to not report a profile when the profiler
stopped with a failure. This seems like a prudent way to go, since the
profile may be in an inconsistent state.

(Out of curiosity, both 1 or 2 would suffice to avoid the issue, but
I think both seem useful individually to avoid future issues.)