Do not try to flush profile when profiler failed
When the worker component fails, we also stop the scheduler. But, up
until now, if there was enough data to be reported, we would still try
to serialize and report a last profile.

This is not an entirely great idea, as the profile may not be in a
consistent state. So, instead, this commit adds a flag to mark this
situation and the scheduler will use it to skip any attempts to
report data.