[PROF-10125] Track unscaled allocation counts in allocation profiler
**What does this PR do?**

This PR extends the allocation profiler to also track the unscaled
allocation counts.

Specifically, like other profile types, the allocation profiler assigns
a weight to every sample, making it "represent" all the objects that
weren't sampled.

**Motivation:**

For debugging, in corner cases, it comes in handy to know exactly how
many samples the profiler observed, and what was the impact of scaling.

As part of preparing the allocation profiler feature for GA, I'm adding
this feature so we can use it to confirm the exact sample counts
whenever needed.

Note also this is something we do for the cpu/wall-time profiler, where
we track the `cpu_or_wall_samples` as an exact count of how many samples
were taken (and again, without the weight -- which in the case of those
profilers, represents time).

**Additional Notes:**

N/A

**How to test the change?**

This change includes test coverage.