[PROF-3760] Revert "Merge pull request #3760 from DataDog/ivoanjo/prof-10123-placeholder-missing-allocations"
**What does this PR do?**

This PR reverts the change from #3760.

While we still want to add this feature, the implementation in #3760
can cause heap profiler to fail with an exception.

This failure also uncovered a few other potential bugs in our error
handling code.

Thus, to be on the safe side, and because we want to get a new release
out, I'm reverting the whole PR, until I can re-open a fixed version.

**Motivation:**

Remove blockers from the 2.2.0 release.

**Additional Notes:**

This reverts commit 7732142aacd95f78f33c4a7d9e7d0af8067d660d, reversing
changes made to ca006e96e9a5bbb40dee78a481329e42dcd57877.

**How to test the change?**

Validate that CI is still green.