Minor: Simplify time retrieval during worker initialization
By getting the timestamp before allocating any memory, we can safely
rely on the exception-raising behavior of our helper, rather than
needing to handle clearing the state if we get a failure
mid-initialization.

Having a timestamp that's a few ns before is not expected to impact the
discrete dynamic sampler behavior.