[PROF-10198] Add benchmark config with allocation profiling + ddprof
**What does this PR do?**

This PR adds a new benchmarking configuration for testing allocation
profiling with `ddprof` enabled.

(As a reminder, `ddprof` is the Datadog native profiler).

**Motivation:**

The intention is to allow us to use ddprof to profile the Ruby profiler,
so we can identify optimization opportunities and validate any changes
we make.

**Additional Notes:**

N/A

**How to test the change?**

I've tested this PR locally using the `bp-runner` tool.