[PROF-10208] Upgrade to libdatadog 11
**What does this PR do?**

This PR upgrades the datadog gem to use libdatadog 11.

There were some minor tweaks needed in the crashtracker APIs, but
otherwise it's a drop-in upgrade.

**Motivation:**

Staying on the latest libdatadog enables us to pick up the
latest improvements and also makes it easier to develop new features
that need libdatadog changes.

**Additional Notes:**

As usual, I'm opening this PR as a draft as libdatadog 11 is not yet
available on rubygems.org, and I'll come back to re-trigger CI and mark
this as non-draft once it is.

**How to test the change?**

Our existing test coverage includes libdatadog testing, so a green CI
is good here :)