[PROF-10198] Set service/env for GitLab runs in benchmark platform
**What does this PR do?**

This PR sets the `DD_SERVICE` and `DD_ENV` for the gitlab benchmark in
the benchmarking platform.

**Motivation:**

While testing a few changes, I realized these benchmarks were running
with no name, thus ending up with `gitlab-runner` which is not a very
clear name (and makes it hard to know what it is).

To align with the reliability environment (`rp-ruby-gitlab`) I've
explicitly set a matching name.

**Additional Notes:**

N/A

**How to test the change?**

I've tested this with the `bp-runner`.