Lower dynamic sampling rate target to avoid flakiness
This test for the "Skipped Samples" placeholder failed in CI:
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/15605/workflows/5b32e4bb-155e-4a94-bf8b-980599eaafae/jobs/568892/tests#failed-test-0>

I strongly suspect the issue is that the profiler was able to "keep up"
with the Ruby VM and thus never needed to skip samples.

In fact, I can reproduce the same behavior by raising the
`dynamic_sampling_rate_overhead_target_percentage` on my machine.

Thus, to reduce flakiness, I'm lowering the target for the dynamic
sampling rate from 2% to 0.1% which should be tight enough to
consistently cause the profiler to need to skip samples in this test.