Avoid retrieving rb_callable_method_entry_t in ddtrace_rb_profile_frames
As explained in the added comment, for our use this is expected to be
equivalent (and our integration tests that compare our output with the
Ruby stack trace APIs still pass).

This gets us a bit less of work during sampling (but not a lot), but
more importantly it will make it easier to add caching in a later
commit.