Introduce `frame_info` to contain sampled stack frame information
Rather than having separate arrays for each kind of information we want
to get from a stack (VALUE *stack_buffer + int *lines_buffer + bool
*is_ruby_frame), introduce a new `frame_info` struct to contain this
information.

This makes it easy to then add or change members in this struct without
having to think about allocating yet more arrays, and managing their
lifetimes and passing them in as arguments. (Spoiler: We'll be using
this in a commit very soon to introduce caching)

As a side bonus of this change, I noticed that we could just expose the
method name for native methods as an `ID`, rather than handing over
the `rb_callable_method_entry_t` ("cme") as we did before.

This has a few advantages:
* It's effectively a way to inline the work done by the previous call
  to `ddtrace_rb_profile_frame_method_name`. Because we had to
  copy/paste some code from the Ruby sources to support
  `ddtrace_rb_profile_frame_method_name` on legacy Rubies, we no
   longer need that code and can delete it!
* An `ID` is interned/immutable, so it's much easier to later cache
  if we want to. (Although no plans yet...)