Introduce caching for sampled stacks
The intuition here is that in many apps, the same stack (or at least the
bottom of the stack) won't change very often -- e.g. during an HTTP
request we may take a bunch of samples, and all of them will have a
common stack that's webserver + rails routing + endpoint doing work.

While right now this mostly avoids us repeating calls to `calc_lineno`,
(calculating the line number for Ruby frames) one of the most
interesting changes in this PR is the new `same_frame` flag.

This `same_frame` flag allows the caller to know when the stack has
changed, and I'm hoping to later connect this to libdatadog
improvements, e.g. if we store the ids for each stack entry in
libdatadog, we can know if the stack changed or not, and thus
if we can just re-feed the same ids to libdatadog or need to record new
ones.