Directly use VM iseq query functions instead of going through rb_profile_ stuff
Since we now always return an iseq from ddtrace_rb_profile_frames
(it's in the name of the struct element even!), we can directly use the
`rb_iseq_` functions.

Looking at the `rb_profile_` functions what they do is check if they
have an iseq or something else, and when they have an iseq, they call
these functions.

So this allows us to skip the extra tests because we know we have iseqs.

(FYI an "iseq" is an "instruction sequence" -- it's the object that
represents the YARV VM bytecode)