Add workaround for spec that started flaking
This was a really weird one. I don't like handwaving it away, but
debugging any more of this would need me to go really deep into Ruby's
GC to understand where exactly Ruby is deciding to keep this object
alive and I'm not sure such a time investment is worth it for this bug,
so I've put in a workaround for now.