[PROF-10241] Add extra tags to crash reports
**What does this PR do?**

This PR adds the following tags to crashtracker reports:
* language
* severity
* profiler_version

Severity is similar to what we do in Python in
https://github.com/DataDog/dd-trace-py/pull/10199 .

The other two tags seem redundant, but right now if we don't pass
them as tags to libdatadog they don't get picked up correctly.

In the future, this won't be needed by libdatadog, but for now this
makes it easier for us to analyze reported crashes.

**Motivation:**

Improve crash reports.

**Additional Notes:**

N/A

**How to test the change?**

Change includes test coverage.