Fix exception in ruby 3.4 development builds
While using a recent Ruby 3.4 development build (to hunt down memory
leaks using `ruby_memcheck`), one of our tests started failing with

> rb_raise(rb_eRuntimeError, "BUG: rb_native_thread* is null. Is this Ruby running with RUBY_MN_THREADS=1?");

This exception had been added as a "not sure what to do, just in case",
but it looks like that even without MN threads it can happen.

(I suspect we may have found a Ruby thread that's so new that it hasn't
had a native thread assigned).

For now let's instead return 0 -- this value is only used for displaying
the thread id, so it seems better to have missing info (a zero), rather
than stopping profiling.

We can always circle back on this once Ruby 3.4 is stable and if we
see this is happening often.