[NO-TICKET] Fix profiler `inline` helpers not working when optimization disabled
**What does this PR do?**

This PR makes sure that every `inline` helper in the profile is
marked as `static inline`.

As documented in
<https://gcc.gnu.org/onlinedocs/gcc/Inline.html#Inline>:

> When a function is both inline and static, if all calls to the function
> are integrated into the caller, and the function’s address is never
> used, then the function’s own assembler code is never referenced.

this is different from just `inline`:

> When an inline function is not static, then the compiler must assume
> that there may be calls from other source files; since a global
> symbol can be defined only once in any program, the function must not
> be defined in the other source files, so the calls therein cannot
> be integrated. Therefore, a non-static inline function is always
> compiled on its own in the usual fashion.

In practice, this manifested as errors such as

> symbol lookup error: datadog_profiling_native_extension.3.4.0_x86_64-linux.so:
> undefined symbol: monotonic_wall_time_now_ns

when building Ruby or the profiler with optimization disabled (`-O0`).

With this fix, these errors no longer show up when building with
optimization disabled.

**Motivation:**

Avoid runtime failures in the profiler on Ruby builds with
optimization disabled / when debugging the profiler.

**Additional Notes:**

Most of these methods with only `inline` were still unreleased
to customers so hopefully nobody was impacted by this issue.

For consistency, I've also made sure that all methods are
`static inline` as we had a mix of `inline static` as well.

**How to test the change?**

I was able to trigger the issue by running

```
$ DDTRACE_DEBUG=true bundle exec rake clean compile
$ bundle exec rspec spec/datadog/profiling/
```