Avoid relying on port from env
This test was failing for me when running outside of our docker testing
environment because it was assuming the port that we use for testing
inside docker (9126) without actually configuring it.

To avoid any issues, I've also set the port in this test so it works
regardless of the configuration of the environment it's run in.