Extract out ruby version for GVL profiling specs
It occurs to me that I'm sprinkling a lot of conditionals all over the
place, and when we want to support 3.2 we'll need to find all of them
and if we miss some there won't be anything reminding us it's missing.

So by extracting the version to a single location, we'll bump it once
and all specs that need to work for 3.2 will be run.