Refactor: Introduce `ThreadContext.for_testing` constructor
So far in the profiler, we've been avoiding adding defaults to keyword
arguments that are expected to be set in production as, well, those
defaults won't match what happens in production (they'll get
overridden).

This is all fine, but quite annoying when we use our components
piecemeal for benchmarking and testing, since adding any new arguments
means modifying a gazillion places in the codebase to start passing in
that argument.

To solve this, let's introduce a `for_testing` helper method that
provides defaults that we often use for testing, and that then calls
into the regular constructor.