[NO-TICKET] Fix flaky spec in profiler due to race
**What does this PR do?**

This PR fixes a flaky spec introduced by #3929: a spec to check that
an error was raised in a background thread implicitly depended on a race
(that the background thread ran before the rspec thread did) and thus
started failing when the race was lost.

By actually synchronizing with the background thread using the
`on_failure_proc`, we now guarantee that the background thread has
the chance to run as expected.

**Motivation:**

Our goal is to always have zero flaky specs in the profiler!

**Additional Notes:**

Fixes https://github.com/DataDog/ruby-guild/issues/179

**How to test the change?**

Validate that CI is still green.