[PROF-10656] Add extra error checking to heap profiling function
**What does this PR do?**

This PR adds extra error checking to the heap profiler's
`on_committed_object_record_cleanup` function.

We've seen one report of a segmentation fault in this function, and
we've not been able to reproduce the issue, so for now I decided to be
extra careful and add a bunch of extra error checking to this function.

**Motivation:**

Investigate potential crash in this function.

**Additional Notes:**

As a reminder, heap profiling is off by default AND still marked as
alpha. The crash we saw was triggered on Ruby 2.7.5, a version where
we need a specific workaround to support heap profiling, so I'm
wondering if the crash is 2.7 specific or not...

Hopefully we'll find out soon.

**How to test the change?**

Validate that CI is still green.