Improve debug output for GVL profiling spec
I saw the following failure in CI
https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/16646/workflows/5a843db3-b63e-4855-8ce9-432c0405e0bb/jobs/597972

```
Failures:

  1) Datadog::Profiling::Collectors::CpuAndWallTimeWorker#start when main thread is sleeping but a background thread is working when GVL profiling is enabled records Waiting for GVL samples
     Failure/Error: expect(waiting_for_gvl_time).to be_within(5).percent_of(total_time)
       expected 402884285 to be within 5% of 492523454
     # ./spec/datadog/profiling/collectors/cpu_and_wall_time_worker_spec.rb:472:in `block (5 levels) in <top (required)>'
     # ./spec/spec_helper.rb:231:in `block (2 levels) in <top (required)>'
     # ./spec/spec_helper.rb:116:in `block (2 levels) in <top (required)>'
     # /usr/local/bundle/gems/webmock-3.24.0/lib/webmock/rspec.rb:39:in `block (2 levels) in <top (required)>'
     # /usr/local/bundle/gems/rspec-wait-0.0.10/lib/rspec/wait.rb:47:in `block (2 levels) in <top (required)>'
```

and I'm not sure why the time accounting is not as expected.

The difference close to ~100ms seems close to the usual 100ms that Ruby
gives each thread to run on CPU, but I'm not quite sure how it would
factor in here.

To be able to further investigate the next time this shows up, I'm
adding extra information that hopefully will point the way.