Clean up configuration printing in error messages
In Ruby 3.4, the printing of arrays is being slightly tweaked, and
so our tests started failing.

Regardless of that change sticking or being reverted (it does seem
like it could cause a bit of annoyance across the ecosystem), I've
decided to simplify a bit our code to provide our own printing rather
than rely on the Ruby defaults, and thus work everywhere.

Also I decided to remove the specs for
`HttpTransport#config_without_api_key` since it's only used
internally for error printing (which did not use to be the case!).