Fix GVL profiling spec still flaking
I was able to debug this last flaking behavior to another race.

The debug output for this one looked like:

```
 :sample_states=>
  ["unknown",
   "had cpu",
   "unknown",
   "unknown",
   "unknown",
   "unknown",
   "unknown",
   "unknown",
   "unknown",
   "unknown",
   "had cpu",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "had cpu",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
   "waiting for gvl",
```

What happened here, is that we missed an entire "Waiting for GVL", but
our logic to categorize missed time in the spec was expecting to
filter only a sequence of unknowns, and got tripped by the "had cpu"
in between them.

So why is that "had cpu" there? It took me a while to figure this out:

What happened is that we saw a tiny amount of cpu, which was the cpu
being used by the thread getting to the "Waiting for GVL" state (e.g.
work the Ruby VM was doing outside of the GVL to put this thread on the
list to run).

The thread was not actually performing useful work, but we don't know
that. Thus, we also need to allow for this "had cpu" in our "missed
time", otherwise we had flaky failures when this situation happened.