Use Qtrue as a placeholder for uninitialized `otel_current_span_key`
As Alex pointed out, using `Qnil` as a marker for uninitialized
and `Qfalse` for failed to initialize is a bit of a footgun since it's
a very subtle detail.

To make it more obvious, I've changed this to `Qtrue` for uninitialized
and `Qnil` for failed to initialize, which hopefully is way harder
to confuse.