[NO-TICKET] Bump debase-ruby_core_source dependency to 3.3.6
**What does this PR do?**

This PR bumps our dependency on debase-ruby_core_source to version 3.3.6. This version includes the headers for the Ruby 3.3.5 as well as 3.4.0-preview2; up until now we were supporting those Ruby versions using the 3.3.0 headers.

**Motivation:**

Use up-to-date headers for profiling on Ruby 3.3.5 and 3.4.0-preview2.

**Additional Notes:**

I've updated the appraisal gemfiles as well.

**How to test the change?**

Validate that CI is green.