[PROF-10589] Add microbenchmark for GVL profiling sampling
**What does this PR do?**

This PR adds a new profiler microbenchmark that tests the performance
of sampling as triggered from the GVL profiling feature.

**Motivation:**

I've created this benchmark as part of validating the performance of
this feature.

**Additional Notes:**

Profiling the profiler with [ddprof](https://github.com/datadog/ddprof),
it looks like right now the biggest performance bottleneck is the cost
of stack sampling.

This can be validated easily -- changing the depth of the stack
for the testing thread makes a big difference.

**How to test the change?**

Validate that this benchmark gets picked up by our usual harness to run
microbenchmarks.